refactor: extract on_vps_ready + compute_fast_endpoint_updates to delivery_live_edge module (#189)
delivery.rs reached 1050 lines after the T13 addition, violating the
project's <1000 line cap. Extract the two new items into a sibling
module crates/rs-api/src/delivery_live_edge.rs:

- compute_fast_endpoint_updates (pure DB helper)
- on_vps_ready (standalone async fn, no longer a method on
  DeliveryOrchestrator — takes the audit_tx as a parameter)

Behavior unchanged. Tests in on_vps_ready_tests.rs updated to the
new module path. The single call site inside start_delivery now
calls the standalone fn with self.pool, self.audit_tx.as_ref().
